상점에 아이템추가, 챔피언레벨표기, 델타타임치트 F6 2배F7 3배, 미니언웨이브 대포번갈아서, 시간마다 돈오름, 유닛죽이면 돈오름

+++ D3DFramework/D3DFramework/PlayerInfoPanel.cpp
@@ -384,7 +384,7 @@ void PlayerInfoPanel::Update()
 	itemshopBtn->SetText((int)champion->stat->GetValue(StatType::Gold));
 
 	// Level
-	levelLabel->SetText(L"%d", champion->stat->GetValue(StatType::Level));
+	levelLabel->SetText(L"%d", (int)champion->stat->GetValue(StatType::Level));
 
 	// exp
 	expBar->uvRatioStart.y = 1.f - (champion->stat->GetValue(StatType::Experience) /
